DOMCore/tests/submissions/Ms2ger/Node-isEqualNode.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<head>
<title>Node.isEqualNode</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"/>
<script>
function testNullHandling(node) {
  test(function() {
    assert_false(node.isEqualNode(null))
    assert_false(node.isEqualNode(undefined))
  })
}
[
  document.createElement("foo"),
  document.createTextNode("foo"),
  document.createProcessingInstruction("foo", "bar"),
  document.createComment("foo"),
  document,
  document.implementation.createDocumentType("html", "", ""),
  document.createDocumentFragment()
].forEach(testNullHandling)

test(function() {
  var a = document.createElement("foo")
  a.setAttribute("a", "bar")
  a.setAttribute("b", "baz")
  var b = document.createElement("foo")
  b.setAttribute("b", "baz")
  b.setAttribute("a", "bar")
  assert_true(a.isEqualNode(b))
}, "isEqualNode should return true when the attributes are in a different order")

test(function() {
  var a = document.createElement("foo")
  a.setAttributeNS("ns", "x:a", "bar")
  var b = document.createElement("foo")
  b.setAttributeNS("ns", "y:a", "bar")
  assert_true(a.isEqualNode(b))
}, "isEqualNode should return true when the attributes have different prefixes")
</script>
</body>
</html>
